adding expcetion handlers for no found replies and notfound  threads

+++ src/main/java/com/api/API2/service/ThreadService.java
@@ -1,6 +1,8 @@
 package com.api.API2.service;
 
 import com.api.API2.entity.Thread;
+import com.api.API2.exceptions.ReplyException;
+import com.api.API2.exceptions.ThreadException;
 import com.api.API2.repository.ThreadRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,12 +24,22 @@ public class ThreadService {
          return threadRepository.save(t);
     }
     public List<Thread> getAllThreads(){
+        List<Thread> threads = threadRepository.findAll();
+        if(!threads.isEmpty())
         return threadRepository.findAll();
+        else throw  new ThreadException("not found");
     }
     public Optional<Thread> getThreadById(String id){
-        return threadRepository.findById(id);
+        Optional<Thread> t= threadRepository.findById(id);
+        if(!t.isEmpty())
+        return t;
+        else throw new ThreadException("not found");
     }
     public void deleteThread(String id){
+
+        Optional<Thread> t= threadRepository.findById(id);
+        if(t.isEmpty())
+            throw new ThreadException("not found");
         threadRepository.deleteById(id);
     }
